add component tracking order
